requirements.txt
gradio>=4.0.0
pinecone-client>=3.0.0
sentence-transformers>=2.2.2
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
aiofiles>=23.1.0
httpx>=0.24.0
PyPDF2>=3.0.0

# Testing and code quality tools
pytest>=7.0
pytest-asyncio>=0.21
mypy>=1.0
black>=23.0
flake8>=6.0
